Fixed incorrect grade returned in getAllMissionProgress
seperated logic that queries question grades to calculate final grade and create task submission. Used resolver chaining so that the function is called everywhere its used.
